[refactor] -Refacto de structure des fichiers notamment les composants et fichiers scss des pages, les composants sont maintenant mieux organises ainsi que les fichiers de style des pages - Decoupage de createPost et createSpace les formulaires on ete deplaces dans createPostForm et createSpaceForm dans components/forms qui vient remplacer le dossier auth - Rework du formulaire de creation d'un espace, ca rend bien et plus propre en code mais plus d'animation de slide a voir ce que l'on peut faire - Fix boutton delete d'un post pas accessible - Ajout d'un cursor pointer sur le boutton delete d'un post et sur la div permetant d'afficher les commentaires - Il y a maintenant un composant Comment qui est importe dans postCard, il s'occuper d'afficher les commentaires et le formulaire
