+++ src/pages/spacePages/CreatePost.js
@@ -1,26 +1,13 @@
-import React, { useState, useRef, useEffect, createRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Redirect } from 'react-router-dom';
-import moment from 'moment';
-import { useForm, toFormData } from '../../utils/forms';
-import PostService from '../../services/PostService';
 import SpaceService from '../../services/SpaceService';
-import UploadInput from '../../components/utilsTemplates/UploadInput/UploadInput';
 import UserAccountHeaederBg from '../../components/userAccount/UserAccountHeaederBg';
 import '../../styles/layout/container.scss';
-import './createPost.scss';
-import NavBar from '../../components/nav/NavBar';
-// import NavBarPrevStep from '../../components/nav/NavBarPrevStep';
+import '../../styles/pages/createPost.scss';
+import CreatePostForm from '../../components/forms/createPostForm/CreatePostForm';
 
 function CreatePost() {
-  const [showPostForm, setShowPostForm] = useState({
-    allForm: false,
-    title: false,
-    image: false,
-    video: false,
-    link: false,
-  });
   const [space, setSpace] = useState(null);
-  const [postValues, handlePostChange, deleteFiles] = useForm();
   const spaceId = window.location.href
     .split('/')[4]
     .substring(window.location.href.split('/')[4].lastIndexOf('-') + 1);
@@ -36,13 +23,6 @@ function CreatePost() {
     setSpace(await SpaceService.focusSpace(spaceId));
   }
 
-  async function createPost(event) {
-    event.preventDefault();
-    const data = toFormData(postValues);
-    await PostService.createPost(spaceId, data);
-    setRedirectToSpace(true);
-  }
-
   if (redirectToSpace)
     return (
       <Redirect
@@ -55,162 +35,12 @@ function CreatePost() {
       <UserAccountHeaederBg />
       <main className="main__createPost">
         <h2>Ajouter un souvenir</h2>
-        <form id="createMemory" action="/" method="post" onSubmit={createPost}>
-          {showPostForm.title && (
-            <div className="input input--title">
-              <label className="input__label" htmlFor="title">
-                Titre du souvenir
-              </label>
-              <input
-                className="input__field"
-                type="text"
-                name="title"
-                id="title"
-                placeholder="Donnez un titre..."
-                value={postValues.title || ''}
-                onChange={handlePostChange}
-              />
-            </div>
-          )}
-          <div className="input input--souvenir">
-            <label className="input__label" htmlFor="content">
-              Votre souvenir
-            </label>
-            <textarea
-              className="input__field"
-              name="text"
-              id="content"
-              cols="30"
-              rows="5"
-              placeholder="Quel souvenir voulez-vous partager ?"
-              value={postValues.text || ''}
-              onChange={handlePostChange}
-              required
-            />
-          </div>
-
-          {showPostForm.link && (
-            <div className="input">
-              <label className="input__label" htmlFor="link">
-                Lien vers une page externe
-              </label>
-              <input
-                className="input__field"
-                type="text"
-                id="link"
-                name="link"
-                placeholder="Ajouter un lien..."
-                value={postValues.link || ''}
-                onChange={handlePostChange}
-              />
-            </div>
-          )}
-          {showPostForm.image && (
-            <div className="input">
-              <div className="input__field">
-                <UploadInput
-                  labelText="Vos images"
-                  handleChange={(e) => {
-                    handlePostChange(e);
-                  }}
-                  isMultiple
-                  images={postValues.imagesFiles}
-                />
-                <div className="previews">
-                  {postValues.imagesFiles.map((image, index) => (
-                    <div className="previews__preview" key={image.name}>
-                      <img
-                        src={URL.createObjectURL(image)}
-                        alt="importées depuis l'ordinateur"
-                      />
-                      <button
-                        type="button"
-                        className="preview__button"
-                        onClick={() => {
-                          deleteFiles(index);
-                        }}
-                      ></button>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
-          )}
-          {showPostForm.video && (
-            <div className="input">
-              <div className="input__field">
-                <UploadInput
-                  labelText="Votre vidéo"
-                  specificFieldName="postVideo"
-                  handleChange={handlePostChange}
-                />
-              </div>
-            </div>
-          )}
-          <div className="buttons">
-            <button
-              type="button"
-              className="button button--add"
-              style={{ display: showPostForm.title ? 'none' : 'inline-block' }}
-              onClick={() => {
-                setShowPostForm(() => {
-                  return { ...showPostForm, title: true };
-                });
-              }}
-            >
-              Ajouter un titre
-            </button>
-            <button
-              type="button"
-              className="button button--add"
-              style={{ display: showPostForm.link ? 'none' : 'inline-block' }}
-              onClick={() =>
-                setShowPostForm(() => {
-                  return { ...showPostForm, link: true };
-                })
-              }
-            >
-              Ajouter un lien externe
-            </button>
-            <button
-              type="button"
-              className="button button--add"
-              style={{ display: showPostForm.image ? 'none' : 'inline-block' }}
-              onClick={() =>
-                setShowPostForm(() => {
-                  return { ...showPostForm, image: true };
-                })
-              }
-            >
-              Ajouter des images
-            </button>
-            <button
-              type="button"
-              className="button button--add"
-              style={{ display: showPostForm.video ? 'none' : 'inline-block' }}
-              onClick={() =>
-                setShowPostForm(() => {
-                  return { ...showPostForm, video: true };
-                })
-              }
-            >
-              Ajouter une video
-            </button>
-          </div>
-          <div className="switchGroup">
-            <div
-              onClick={() => setToggleBtn(!toggleBtn)}
-              className={`toggle-btn ${toggleBtn ? 'active' : ''}`}
-            >
-              <input type="checkbox" checked className="cb-value" />
-              <span className="round-btn"></span>
-            </div>
-            <p className="text">
-              Je suis sur que le défunt serait d'accord pour que je publie cela.
-              Aussi, je suis sur que personne ne serait offensé par ce contenu.
-            </p>
-          </div>
-        </form>
+        <CreatePostForm
+          spaceId={spaceId}
+          setRedirectToSpace={setRedirectToSpace}
+          toggleBtn={toggleBtn}
+          setToggleBtn={setToggleBtn}
+        />
       </main>
       <footer className={`footer ${toggleBtn ? 'is-visible' : ''}`}>
         <div className="footer--border">
@@ -235,9 +65,9 @@ function CreatePost() {
                 width="328"
                 height="60"
                 filterUnits="userSpaceOnUse"
-                color-interpolation-filters="sRGB"
+                colorInterpolationFilters="sRGB"
               >
-                <feFlood flood-opacity="0" result="BackgroundImageFix" />
+                <feFlood floodOpacity="0" result="BackgroundImageFix" />
                 <feColorMatrix
                   in="SourceAlpha"
                   type="matrix"
