+++ src/pages/userAccountPages/UserAccountPage.js
@@ -1,13 +1,13 @@
 import React, { useState, useContext, useEffect, useRef } from 'react';
 import { Link } from 'react-router-dom';
 import UserContext from '../../context/UserContext';
-import SpacesContainer from '../../components/userAccount/SpacesContainer';
+import SpacesContainer from '../../components/userAccount/spacesContainer/SpacesContainer';
 import UserInvitationContainer from '../../components/userAccount/UserInvitationsContainer';
-import RequestAccessContainer from '../../components/utilsTemplates/requestAccessContainer/RequestAccessContainer';
+import RequestAccessContainer from '../../components/userAccount/RequestAccessContainer';
 import UserAccountHeaederBg from '../../components/userAccount/UserAccountHeaederBg';
 // import logoPuzzle from '../../assets/svg/puzzle-logo.svg';
 import '../../components/nav/navbar-userAccount.scss';
-import './userAccount.scss';
+import '../../styles/pages/userAccount.scss';
 import '../../styles/layout/nav.scss';
 import useWindowSize from '../../utils/useWindowSize';
 
@@ -49,44 +49,31 @@ function UserAccountPage() {
     // width > 750 ? seTranslateValue(750) : seTranslateValue(100);
   }
 
-  // console.log('hez', translateValue);
-
   function handleClickNav(id) {
     const wrapper = document.querySelector('.wrapper--flex');
+    width < 750
+      ? (wrapper.style.transform = `translateX(${id * -100}vw)`)
+      : (wrapper.style.transform = `translateX(${id * -750}px)`);
     switch (id) {
       case 0:
         setIsTab1Active(true);
         setIsTab2Active(false);
         setIsTab3Active(false);
-        width < 750
-          ? (wrapper.style.transform = `translateX(${id * -100}vw)`)
-          : (wrapper.style.transform = `translateX(${id * -750}px)`);
         break;
       case 1:
         setIsTab1Active(false);
         setIsTab2Active(true);
         setIsTab3Active(false);
-        // wrapper.style.transform = `translateX(${id * -100}vw)`;
-        width < 750
-          ? (wrapper.style.transform = `translateX(${id * -100}vw)`)
-          : (wrapper.style.transform = `translateX(${id * -750}px)`);
         break;
       case 2:
         setIsTab1Active(false);
         setIsTab2Active(false);
         setIsTab3Active(true);
-        // wrapper.style.transform = `translateX(${id * -100}vw)`;
-        width < 750
-          ? (wrapper.style.transform = `translateX(${id * -100}vw)`)
-          : (wrapper.style.transform = `translateX(${id * -750}px)`);
         break;
       default:
         setIsTab1Active(true);
         setIsTab2Active(false);
         setIsTab3Active(false);
-        width < 750
-          ? (wrapper.style.transform = `translateX(${id * -100}vw)`)
-          : (wrapper.style.transform = `translateX(${id * -750}px)`);
         break;
     }
   }
