+++ src/styles/config/_variables.scss
@@ -20,3 +20,8 @@ $roboto-bold: 'Roboto Bold';
     @content;
   }
 }
+@mixin desktop {
+  @media (min-width: 1024px) {
+    @content;
+  }
+}
