+++ src/styles/pages/_loginPage.scss
@@ -1,8 +1,18 @@
+@import '../config/variables';
+
 .login {
   .header {
     height: 40vh;
   }
   .header__img {
     background-position: center bottom 70%;
+  }
+}
+
+@include min-sm {
+  .login {
+    .wrapper--flex {
+      justify-content: center;
+    }
   }
 }
