+++ src/styles/pages/_space.scss
@@ -223,6 +223,7 @@
     display: flex;
     justify-content: space-between;
     width: 100%;
+    cursor: pointer;
     p {
       font-weight: 300;
       font-size: 10px;
@@ -283,6 +284,7 @@
     display: flex;
     justify-content: space-between;
     padding-top: 15px;
+    align-items: flex-end;
     .image-preview {
       display: flex;
       flex-direction: column;
