+++ src/styles/pages/spaceSettings.scss
